mindctrl-addon/requirements.txt
# https://raw.githubusercontent.com/mlflow/mlflow/master/requirements/gateway-requirements.txt

mlflow~=2.10
pydantic<3,>=1.0
fastapi<1
uvicorn[standard]<1
watchfiles<1
aiohttp<4
boto3<2,>=1.28.56
openai<1
tiktoken
tenacity
# homeassistant>2023.11 # TODO: I need to switch to Arch from debian :(
# https://aur.archlinux.org/packages/python-mlflow
azureml-mlflow
paho-mqtt
